backend/requirements.txt
fastapi==0.103.1
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy>=1.4.23,<1.5.0
psycopg2-binary>=2.9.1,<2.10.0
alembic>=1.7.1,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart>=0.0.5,<0.0.6
jinja2>=3.0.1,<3.1.0
email-validator>=2.0.0,<2.2.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.21.0 